refactor(p1): extract SandboxActivitySheet Stage 3 — cross-pkg visibility
Move SandboxActivitySheet (50 LOC) from ChatInput.kt (783 → 729) to ChatInputSandbox.kt as `internal fun` (was top-level `fun` = public). Cross-package caller `ChatPage.kt:93 import + L804 call` still resolves because `internal` is module-scoped (both files in app module).

Demote 3 sibling composables now that Sheet is in the same file:
- SandboxSheetHeader: internal → private
- SandboxToolActivityContent: internal → private
- SandboxWebActivityContent: internal → private
(SandboxPeekBar stays internal — still called by main ChatInput in old file.)

Drop 3 imports from ChatInput.kt that became orphan by this move: ModalBottomSheet, rememberModalBottomSheetState, fillMaxHeight. BackHandler kept (still 1 caller in old file). Pre-existing dead `WorkspaceTone` import preserved per CLAUDE.md §3.

Byte-equal Sheet body modulo visibility modifier. Baseline 6/427 unit tests preserved.

ChatInput.kt: 1811 → 729 (-60%) — Sandbox extraction complete.
ChatInputSandbox.kt: 1135 lines (helpers + 14 composables + Sheet).
